chore: disable macaroons
by disabling the macaroons feature we cannot listen on 0.0.0.0 anymore because LND forbids it. Hence, we only listen on our local container IP.
By doing so we do not randomly generate new macaroons and our http api is accessible by anyone.
We had to disable grpc because it is not allowed to have it without macaroons. This is not a problem because we do not make use of this interface anyways.